+++ src/Api/Providers/Auth.php
@@ -155,7 +155,7 @@ class Auth extends Provider
     protected function processLogin($username, $password)
     {
         $this->request->clearToken();
-        $this->request->httpClient()->removeCookies();
+        $this->request->getHttpClient()->removeCookies();
 
         $credentials = [
             'username_or_email' => $username,